fix: prevent double encode if we are on /search endpoint
